% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_or.R
\name{single_or}
\alias{single_or}
\title{Single Odds Ratio Comparator}
\usage{
single_or(data, case_cohort, control_cohort, comparator, eval = TRUE)
}
\arguments{
\item{data}{Dataframe containing all necessary data for comparison}

\item{case_cohort}{The first group to be compared. Can be formulaic (see "eval"), \code{numeric} or \code{character} based}

\item{control_cohort}{The second group to be compared. Can be formulaic (see "eval"), \code{numeric} or \code{character} based}

\item{comparator}{How the groups are compared. These are \code{character} vectors.}

\item{eval}{Default TRUE. When Eval flag is \code{TRUE}, cohorts will be based on formulaic filtering.  When \code{FALSE}, cohort will be based on label filtering}
}
\value{
Returns a \code{dataframe} of Odds Ratios and summary data per cohort
}
\description{
Generate a Fisher's Exact Test summary table based on a single comparator between two cohorts in a single dataframe
}
\examples{
single_or(
  data = iris,
  case_cohort = iris$Petal.Length > 1.2,
  control_cohort = iris$Petal.Length <= 1.2,
  comparator = Species,
  eval = TRUE
)

}
\seealso{
\code{\link{fisher.test}}
}
