% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_or.R
\name{single_or}
\alias{single_or}
\title{Single Odds Ratio Comparator}
\usage{
single_or(data, case_cohort, control_cohort, comparator)
}
\arguments{
\item{data}{A \code{data.frame} containing all necessary data for comparison.}

\item{case_cohort}{The first group to be compared from \code{data}. This should be an expression.}

\item{control_cohort}{The second group to be compared from \code{data}. This should be an expression.}

\item{comparator}{How the groups are compared. This variable should reference a column within \code{data}.}
}
\value{
Returns a \code{dataframe} of Odds Ratios and summary data per cohort
}
\description{
Generate a Fisher's Exact Test summary table based on a single comparator between two cohorts in a single dataframe
}
\examples{
single_or(
  data = iris,
  case_cohort = Petal.Length > 1.2,
  control_cohort = Petal.Length <= 1.2,
  comparator = Species
)

}
\seealso{
\code{\link{fisher.test}}
}
